refactor: Move classes from api to impl to reduce api surface
