+++ testing/camunda-process-test-langchain4j/src/main/java/io/camunda/process/test/impl/judge/BedrockChatModelBuilder.java
@@ -13,13 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.process.test.api.judge;
+package io.camunda.process.test.impl.judge;
 
-import static io.camunda.process.test.api.judge.ModelBuilderSupport.hasText;
-import static io.camunda.process.test.api.judge.ModelBuilderSupport.require;
+import static io.camunda.process.test.impl.judge.ModelBuilderSupport.hasText;
+import static io.camunda.process.test.impl.judge.ModelBuilderSupport.require;
 
 import dev.langchain4j.model.bedrock.BedrockChatModel;
 import dev.langchain4j.model.chat.ChatModel;
+import io.camunda.process.test.api.judge.JudgeConfigBootstrapData;
 import java.util.List;
 import java.util.Map;
 import org.slf4j.Logger;
