+++ testing/camunda-process-test-langchain4j/src/main/java/io/camunda/process/test/impl/judge/ModelBuilderSupport.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.process.test.api.judge;
+package io.camunda.process.test.impl.judge;
 
 final class ModelBuilderSupport {
