+++ testing/camunda-process-test-langchain4j/src/main/resources/META-INF/services/io.camunda.process.test.api.judge.JudgeConfigBootstrapProvider
@@ -1,1 +1,1 @@
-io.camunda.process.test.api.judge.JudgeConfigBootstrap
+io.camunda.process.test.impl.judge.JudgeConfigBootstrap
