+++ testing/camunda-process-test-spring/src/test/java/io/camunda/process/test/api/JudgeConfigBootstrapIT.java
@@ -20,7 +20,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.camunda.process.test.api.judge.ChatModelAdapter;
 import io.camunda.process.test.api.judge.JudgeConfig;
-import io.camunda.process.test.api.judge.JudgeConfigBootstrap;
+import io.camunda.process.test.impl.judge.JudgeConfigBootstrap;
 import io.camunda.process.test.api.judge.JudgeConfigBootstrapData;
 import io.camunda.process.test.api.judge.JudgeConfigBootstrapData.OpenAiConfig;
 import io.camunda.process.test.api.judge.JudgeConfigBootstrapData.ProviderConfig;
